fix(output): export output writer
Termenv output can be a buffer, ssh session, a file, or anything that
implements the io.Writer interface.

In the case of an ssh session, the std ssh library returns a
io.ReadWriter for ssh sessions and the current Termenv implementation
makes it impossible to read terminal status reports since it expects a
file in return.

In addition, the TTY() function is confusing since it implies that the
returned value is a TTY. Which is not always true until we check that
using the isTTY() function.

Deprecate TTY() in favor of Writer() which simply returns the underlying
writer. The caller then can infer the type of the writer and decide what
to do with it.

This also reverts parts of commit 669c9abfb65169a7a146968e38702d3d80327a42 https://github.com/muesli/termenv/commit/669c9abfb65169a7a146968e38702d3d80327a42
Specifically, it removes the File interface.