fix(output): report terminal status when writer is not a file
The underlying writer doesn't have to be a *os.File for it to be a TTY.
For example, a PTY ssh session is a TTY. However, the std library
returns a io.ReadWriter for the ssh session.

Combined with the WithUnsafe() option, we can query the terminal of an
ssh session using Termenv.