fix(output): cache bg and fg colors by default
It's uncommon for users to change the colorscheme of their terminal from
a dark to light, or vice versa, while a program is running. This makes
Termenv query the terminal for colors too many times.

Enable color caching by default and give the program the option to
opt-out using WithColorCache(false).